+++ cypress/locators/date-input/locators.js
@@ -1,6 +1,4 @@
 // component preview locators
-export const DATE_INPUT = '[data-element="input"]';
-export const DATE_ICON = 'span[data-component="icon"][data-element="calendar"]';
-export const DAY_PICKER_WRAPPER = 'div[class="DayPicker-wrapper"]';
-export const DAY_PICKER_HEADING = ".DayPicker-Caption";
-export const DAY = ".DayPicker-Day";
+const DATE_INPUT = '[data-element="input"]';
+
+export default DATE_INPUT;