+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+## [124.2.0](https://github.com/Sage/carbon/compare/v124.1.0...v124.2.0) (2023-12-06)
+
+
+### Features
+
+* **button-toggle:** add minor child ([864f005](https://github.com/Sage/carbon/commit/864f005fd608514390a5cfb2d89f2714826b0b53))
+
+## [124.1.0](https://github.com/Sage/carbon/compare/v124.0.0...v124.1.0) (2023-12-04)
+
+
+### Features
+
+* **dialog:** add greyBackground prop ([416b0cf](https://github.com/Sage/carbon/commit/416b0cf5bdc475ac3f1d5a0e48143b865c0a712e)), closes [#6257](https://github.com/Sage/carbon/issues/6257)
+
+## [124.0.0](https://github.com/Sage/carbon/compare/v123.11.2...v124.0.0) (2023-12-01)
+
+
+### ⚠ BREAKING CHANGES
+
+* **show-edit-pod:** `ShowEditPod` no longer exists
+* **menu-segment-title:** You will need to pass any associated list items as children of the
+`MenuSegmentTitle` so they are rendered in a sublist with the segment heading
+
+### Bug Fixes
+
+* **menu-item:** ensure that icons rendered in menu item have display inline-block ([a937011](https://github.com/Sage/carbon/commit/a93701171e9f2931ddba25eee09d3681557fa7fc))
+* **menu-segment-title:** ensure segment titles are rendered as part of sublist with associated items ([9f4a6ff](https://github.com/Sage/carbon/commit/9f4a6ffd91c9cf6b8a308b11362ebf7671adfd46)), closes [#6010](https://github.com/Sage/carbon/issues/6010)
+
+
+### Miscellaneous Chores
+
+* **show-edit-pod:** deprecate component ([3183e6c](https://github.com/Sage/carbon/commit/3183e6c2b046b8894e6f190fcb88b2bb1c54b744))
+
 ### [123.11.2](https://github.com/Sage/carbon/compare/v123.11.1...v123.11.2) (2023-11-27)